http: make self->curl actually a CURL pointer instead of gchar *
Seemingly the CURL typedef is actually void, so any kind of pointer
is acceptable to curl functions, not causing warnings even if I change
it to a gint pointer. I kind of don't like this, and as it seems
if I define CURL_NO_OLDIES we get a proper typedef.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>