http: don't expect 100-continue headers
Whenever the POST data exceeds 1024 bytes, curl will automatically
emit an "Expect: 100-continue" header, which in turn would cause a delay
on the POST, for two reasons:

1) even if the server implements 100 Continue, it's an extra round trip
2) if the server doesn't, curl waits up to 1sec for this header.

My expectation is that most of the special case servers that collectors
use, doesn't even support this, it makes sense to disable it. Also,
these collectors should be ready to accept >1024 POST bodies, especially
if they support batching.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>