logthrdestdrv: move batch size increment before calling insert
To avoid an off-by one when checking if flush needs to be done, increment
batch_size before calling insert. That's also much better from pinciple,
the current batch implies the actual message being inserted, so if
we return SUCCESS, the entire batch, including the current message would
be assumed to have been delivered.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>